Переместил папочки в нужное место
