+++ app/src/main/java/dev/rodosteam/questtime/screen/editor/EditorFragment.kt
@@ -1,20 +1,19 @@
-package dev.rodosteam.questtime.screen.common.settings
+package dev.rodosteam.questtime.screen.editor
 
 import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import android.widget.TextView
-import androidx.fragment.app.Fragment
 import androidx.lifecycle.Observer
 import androidx.lifecycle.ViewModelProvider
-import dev.rodosteam.questtime.databinding.FragmentSettingsBinding
+import dev.rodosteam.questtime.databinding.FragmentEditorBinding
 import dev.rodosteam.questtime.screen.common.base.BaseFragment
 
-class SettingsFragment : BaseFragment() {
+class EditorFragment : BaseFragment() {
 
-    private lateinit var settingsViewModel: SettingsViewModel
-    private var _binding: FragmentSettingsBinding? = null
+    private lateinit var editorViewModel: EditorViewModel
+    private var _binding: FragmentEditorBinding? = null
 
     // This property is only valid between onCreateView and
     // onDestroyView.
@@ -25,14 +24,14 @@ class SettingsFragment : BaseFragment() {
         container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
-        settingsViewModel =
-            ViewModelProvider(this).get(SettingsViewModel::class.java)
+        editorViewModel =
+            ViewModelProvider(this).get(EditorViewModel::class.java)
 
-        _binding = FragmentSettingsBinding.inflate(inflater, container, false)
+        _binding = FragmentEditorBinding.inflate(inflater, container, false)
         val root: View = binding.root
 
-        val textView: TextView = binding.textSettings
-        settingsViewModel.text.observe(viewLifecycleOwner, Observer {
+        val textView: TextView = binding.textEditor
+        editorViewModel.text.observe(viewLifecycleOwner, Observer {
             textView.text = it
         })
         return root
