+++ app/src/main/java/dev/rodosteam/questtime/screen/external/ExternalViewModel.kt
@@ -1,13 +1,13 @@
-package dev.rodosteam.questtime.screen.common.settings
+package dev.rodosteam.questtime.screen.external
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 
-class SettingsViewModel : ViewModel() {
+class ExternalViewModel : ViewModel() {
 
     private val _text = MutableLiveData<String>().apply {
-        value = "WIP App Settings"
+        value = "This is external Fragment"
     }
     val text: LiveData<String> = _text
 }
