+++ app/src/main/java/dev/rodosteam/questtime/screen/settings/SettingsFragment.kt
@@ -1,20 +1,19 @@
-package dev.rodosteam.questtime.screen.common.external
+package dev.rodosteam.questtime.screen.settings
 
 import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import android.widget.TextView
-import androidx.fragment.app.Fragment
 import androidx.lifecycle.Observer
 import androidx.lifecycle.ViewModelProvider
-import dev.rodosteam.questtime.databinding.FragmentExternalBinding
+import dev.rodosteam.questtime.databinding.FragmentSettingsBinding
 import dev.rodosteam.questtime.screen.common.base.BaseFragment
 
-class ExternalFragment : BaseFragment() {
+class SettingsFragment : BaseFragment() {
 
-    private lateinit var externalViewModel: ExternalViewModel
-    private var _binding: FragmentExternalBinding? = null
+    private lateinit var settingsViewModel: SettingsViewModel
+    private var _binding: FragmentSettingsBinding? = null
 
     // This property is only valid between onCreateView and
     // onDestroyView.
@@ -25,12 +24,16 @@ class ExternalFragment : BaseFragment() {
         container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
-        externalViewModel =
-            ViewModelProvider(this).get(ExternalViewModel::class.java)
+        settingsViewModel =
+            ViewModelProvider(this).get(SettingsViewModel::class.java)
 
-        _binding = FragmentExternalBinding.inflate(inflater, container, false)
+        _binding = FragmentSettingsBinding.inflate(inflater, container, false)
         val root: View = binding.root
 
+        val textView: TextView = binding.textSettings
+        settingsViewModel.text.observe(viewLifecycleOwner, Observer {
+            textView.text = it
+        })
         return root
     }
