+++ app/src/main/java/dev/rodosteam/questtime/screen/settings/SettingsViewModel.kt
@@ -1,13 +1,13 @@
-package dev.rodosteam.questtime.screen.common.external
+package dev.rodosteam.questtime.screen.settings
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 
-class ExternalViewModel : ViewModel() {
+class SettingsViewModel : ViewModel() {
 
     private val _text = MutableLiveData<String>().apply {
-        value = "This is external Fragment"
+        value = "WIP App Settings"
     }
     val text: LiveData<String> = _text
 }
