+++ app/src/main/res/layout/fragment_editor.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".screen.common.editor.EditorFragment">
+    tools:context=".screen.editor.EditorFragment">
 
     <TextView
         android:id="@+id/text_editor"
