+++ app/src/main/res/layout/fragment_external.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".screen.common.external.ExternalFragment">
+    tools:context=".screen.external.ExternalFragment">
 
     <androidx.recyclerview.widget.RecyclerView
         android:id="@+id/external_recycler_view"
