+++ app/src/main/res/layout/fragment_library.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".screen.common.library.LibraryFragment">
+    tools:context=".screen.library.LibraryFragment">
 
     <androidx.recyclerview.widget.RecyclerView
         android:id="@+id/library_recycler_view"
