+++ app/src/main/res/layout/fragment_settings.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".screen.common.settings.SettingsFragment">
+    tools:context=".screen.settings.SettingsFragment">
 
     <TextView
         android:id="@+id/text_settings"
